Back-end/requirements.txt
# FastAPI Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Redis Cache & Background Tasks
redis==5.0.1
hiredis==2.3.2
celery==5.3.6

# HTTP Client
httpx==0.26.0

# Firebase
firebase-admin==6.4.0

# IoT Protocols
paho-mqtt==1.6.1
websockets==12.0

# AI/ML
scikit-learn==1.4.0
numpy==1.26.3
joblib==1.3.2
scipy==1.11.4

# File Upload
cloudinary==1.38.0

# WebRTC Signaling
python-socketio==5.11.0

# Utilities
python-dotenv==1.0.0
